Split Graphics/Mesh module from Graphics/Primitives
