+++ Apps/Samples/Asteroids/Asteroid.h
@@ -25,7 +25,7 @@ Random generated asteroid model with mesh and texture ready for rendering.
 
 #include <Methane/Graphics/RenderContext.h>
 #include <Methane/Graphics/MeshBuffers.hpp>
-#include <Methane/Graphics/Mesh/IcosahedronMesh.hpp>
+#include <Methane/Graphics/IcosahedronMesh.hpp>
 
 namespace hlslpp // NOSONAR
 {
