+++ Apps/Tutorials/03-TexturedCube/TexturedCubeApp.cpp
@@ -24,7 +24,7 @@ Tutorial demonstrating textured cube rendering with Methane graphics API
 #include "TexturedCubeApp.h"
 
 #include <Methane/Samples/AppSettings.hpp>
-#include <Methane/Graphics/Mesh/CubeMesh.hpp>
+#include <Methane/Graphics/CubeMesh.hpp>
 #include <Methane/Data/TimeAnimation.h>
 
 #include <magic_enum.hpp>
