+++ Modules/Graphics/CMakeLists.txt
@@ -2,5 +2,6 @@ add_subdirectory(App)
 add_subdirectory(Core)
 add_subdirectory(Extensions)
 add_subdirectory(Primitives)
+add_subdirectory(Mesh)
 add_subdirectory(Camera)
 add_subdirectory(Types)
