+++ Modules/Graphics/Mesh/Include/Methane/Graphics/BaseMesh.hpp
