+++ Modules/Graphics/Mesh/Include/Methane/Graphics/CubeMesh.hpp
