+++ Modules/Graphics/Mesh/Include/Methane/Graphics/IcosahedronMesh.hpp
