+++ Modules/Graphics/Mesh/Include/Methane/Graphics/QuadMesh.hpp
