+++ Modules/Graphics/Mesh/Include/Methane/Graphics/UberMesh.hpp
